Default

Index

CTFlows.__autonomous

CTFlows.__autonomous

CTFlows.__autonomous

CTFlows.__variable

CTFlows.__variable

Warning

In the examples in the documentation below, the methods are not prefixed by the module name even if they are private.

julia> using CTFlows
julia> x = 1
julia> private_fun(x) # throw an error

must be replaced by

julia> using CTFlows
julia> x = 1
julia> CTFlows.private_fun(x)

However, if the method is reexported by another package, then, there is no need of prefixing.

julia> module OptimalControl
           import CTFlows: private_fun
           export private_fun
       end
julia> using OptimalControl
julia> x = 1
julia> private_fun(x)

Documentation

CTFlows.__autonomous — Method

__autonomous(_::CTModels.Model{CTModels.Autonomous}) -> Bool

Return true for a model declared as CTModels.Autonomous.

Used to determine whether a model has time-independent dynamics.

CTFlows.__autonomous — Method

__autonomous(
    _::CTModels.Model{CTModels.NonAutonomous}
) -> Bool

Return false for a model declared as CTModels.NonAutonomous.

Used to identify models whose dynamics depend explicitly on time.

CTFlows.__autonomous — Method

__autonomous() -> Bool

Return true by default, assuming the problem is autonomous.

This is the fallback for generic cases when no model is provided.

CTFlows.__variable — Method

__variable(ocp::CTModels.Model) -> Bool

Return true if the model has one or more external variables.

Used to check whether the problem is parameterized by an external vector v.

CTFlows.__variable — Method

__variable() -> Bool

Return false by default, assuming no external variable is used.

Fallback for cases where no model is given.