Internal functions

CTBase.OptimalControlSolution — Method

OptimalControlSolution(
    ocfs::CTFlows.OptimalControlFlowSolution
) -> CTBase.OptimalControlSolution

Construct an OptimalControlSolution from an OptimalControlFlowSolution.

CTFlows.ExtensionError — Type

struct ExtensionError <: CTBase.CTException

Exception thrown when an extension is not loaded but the user tries to call a function of it.

Fields

var::String

CTFlows.OptimalControlFlowSolution — Type

struct OptimalControlFlowSolution

Type of an optimal control flow solution.

Fields

ode_sol::Any

feedback_control::CTBase.ControlLaw

ocp::CTBase.OptimalControlModel

variable::Union{Real, AbstractVector{<:Real}}

Base.showerror — Method

showerror(io::IO, e::CTFlows.ExtensionError)

Print the exception.

CTFlows.__abstol — Method

__abstol() -> Float64

Default absolute tolerance for ODE solvers.

See abstol from DifferentialEquations.

CTFlows.__alg — Method

__alg() -> Any

Default algorithm for ODE solvers.

See alg from DifferentialEquations.

CTFlows.__callback — Method

__callback()

See callback from DifferentialEquations.

CTFlows.__reltol — Method

__reltol() -> Float64

Default relative tolerance for ODE solvers.

See reltol from DifferentialEquations.

CTFlows.__saveat — Method

__saveat() -> Vector{Any}

See saveat from DifferentialEquations.

CTFlows.__tstops — Method

__tstops() -> Vector{Real}

See tstops from DifferentialEquations.

CTFlows.makeH — Method

makeH(
    f::CTBase.Dynamics,
    u::CTBase.ControlLaw,
    f⁰::CTBase.Lagrange,
    p⁰::Real,
    s::Real,
    g::CTBase.MixedConstraint,
    μ::CTBase.Multiplier
) -> CTFlows.var"#H#43"{CTBase.Dynamics{time_dependence, variable_dependence}, CTBase.ControlLaw{time_dependence1, variable_dependence1}, CTBase.Lagrange{time_dependence2, variable_dependence2}, var"#s178", var"#s1781", CTBase.MixedConstraint{time_dependence3, variable_dependence3}, CTBase.Multiplier{time_dependence4, variable_dependence4}} where {time_dependence, variable_dependence, time_dependence1, variable_dependence1, time_dependence2, variable_dependence2, var"#s178"<:Real, var"#s1781"<:Real, time_dependence3, variable_dependence3, time_dependence4, variable_dependence4}

Constructs the Hamiltonian:

H(t, x, p) = p ⋅ f(t, x, u(t, x, p)) + s p⁰ f⁰(t, x, u(t, x, p)) + μ(t, x, p) ⋅ g(t, x, u(t, x, p))

CTFlows.makeH — Method

makeH(
    f::CTBase.Dynamics,
    u::CTBase.ControlLaw,
    f⁰::CTBase.Lagrange,
    p⁰::Real,
    s::Real
) -> CTFlows.var"#H#41"{CTBase.Dynamics{time_dependence, variable_dependence}, CTBase.ControlLaw{time_dependence1, variable_dependence1}, CTBase.Lagrange{time_dependence2, variable_dependence2}, <:Real, <:Real} where {time_dependence, variable_dependence, time_dependence1, variable_dependence1, time_dependence2, variable_dependence2}

Constructs the Hamiltonian:

H(t, x, p) = p ⋅ f(t, x, u(t, x, p)) + s p⁰ f⁰(t, x, u(t, x, p))

CTFlows.makeH — Method

makeH(
    f::CTBase.Dynamics,
    u::CTBase.ControlLaw,
    g::CTBase.MixedConstraint,
    μ::CTBase.Multiplier
) -> CTFlows.var"#H#42"{CTBase.Dynamics{time_dependence, variable_dependence}, CTBase.ControlLaw{time_dependence1, variable_dependence1}, CTBase.MixedConstraint{time_dependence2, variable_dependence2}, CTBase.Multiplier{time_dependence3, variable_dependence3}} where {time_dependence, variable_dependence, time_dependence1, variable_dependence1, time_dependence2, variable_dependence2, time_dependence3, variable_dependence3}

Constructs the Hamiltonian:

H(t, x, p) = p ⋅ f(t, x, u(t, x, p)) + μ(t, x, p) ⋅ g(t, x, u(t, x, p))

CTFlows.makeH — Method

makeH(
    f::CTBase.Dynamics,
    u::CTBase.ControlLaw
) -> CTFlows.var"#39#40"{CTBase.Dynamics{time_dependence, variable_dependence}, CTBase.ControlLaw{time_dependence1, variable_dependence1}} where {time_dependence, variable_dependence, time_dependence1, variable_dependence1}

Constructs the Hamiltonian:

H(t, x, p) = p f(t, x, u(t, x, p))