Extension Default

Index

CTFlowsODE.__abstol

CTFlowsODE.__alg

CTFlowsODE.__callback

CTFlowsODE.__internalnorm

CTFlowsODE.__reltol

CTFlowsODE.__saveat

CTFlowsODE.__thevariable

CTFlowsODE.__thevariable

CTFlowsODE.__thevariable

CTFlowsODE.__tstops

Warning

In the examples in the documentation below, the methods are not prefixed by the module name even if they are private.

julia> using CTFlows
julia> x = 1
julia> private_fun(x) # throw an error

must be replaced by

julia> using CTFlows
julia> x = 1
julia> CTFlows.private_fun(x)

However, if the method is reexported by another package, then, there is no need of prefixing.

julia> module OptimalControl
           import CTFlows: private_fun
           export private_fun
       end
julia> using OptimalControl
julia> x = 1
julia> private_fun(x)

Documentation

CTFlowsODE.__abstol — Method

__abstol() -> Float64

Default absolute tolerance for ODE solvers.

See abstol from DifferentialEquations.

CTFlowsODE.__alg — Method

__alg(

) -> Tsit5{typeof(OrdinaryDiffEqCore.trivial_limiter!), typeof(OrdinaryDiffEqCore.trivial_limiter!), Static.False}

Default algorithm for ODE solvers.

See alg from DifferentialEquations.

CTFlowsODE.__callback — Method

__callback()

See callback from DifferentialEquations.

CTFlowsODE.__internalnorm — Method

__internalnorm() -> CTFlowsODE.var"#1#2"

Default internal norm.

CTFlowsODE.__reltol — Method

__reltol() -> Float64

Default relative tolerance for ODE solvers.

See reltol from DifferentialEquations.

CTFlowsODE.__saveat — Method

__saveat() -> Vector{Any}

See saveat from DifferentialEquations.

CTFlowsODE.__thevariable — Method

__thevariable(t0, x0, p0, tf, ocp) -> Any

Default variable from ocp.

CTFlowsODE.__thevariable — Method

__thevariable(x0, p0) -> Vector

Default variable x0, p0.

CTFlowsODE.__thevariable — Method

__thevariable(x0) -> Vector

Default variable from x0.

CTFlowsODE.__tstops — Method

__tstops() -> Vector{Real}

See tstops from DifferentialEquations.